feat(): move getDefaultEntitySettings so it can be exported
