+++ packages/common/src/discussions/api/settings/index.ts
@@ -1,1 +1,2 @@
 export * from "./settings";
+export * from "./getDefaultEntitySettings";
